Made test more generic
