+++ tests/data/resources/dummy_resources.json
@@ -1,0 +1,59 @@
+{
+    "0": {
+        "some_dummy_resources": {
+            "resource_a": {
+                "some_id": 1,
+                "region": "some_region",
+                "some_other_scope": "0"
+            },
+            "resource_b": {
+                "some_id": 2,
+                "region": "some_region",
+                "some_other_scope": "0"
+            }
+        },
+        "some_dummy_resources_count": 2,
+        "other_dummy_resources": {
+            "resource_a": {
+                "some_id": 1,
+                "region": "some_region",
+                "some_other_scope": "0"
+            },
+            "resource_b": {
+                "some_id": 2,
+                "region": "some_region",
+                "some_other_scope": "0"
+            }
+        },
+        "other_dummy_resources_count": 2
+    },
+    "1": {
+        "some_dummy_resources": {
+            "resource_a": {
+                "some_id": 1,
+                "region": "some_region",
+                "some_other_scope": "1"
+            },
+            "resource_b": {
+                "some_id": 2,
+                "region": "some_region",
+                "some_other_scope": "1"
+            }
+        },
+        "some_dummy_resources_count": 2,
+        "other_dummy_resources": {
+            "resource_a": {
+                "some_id": 1,
+                "region": "some_region",
+                "some_other_scope": "1"
+            },
+            "resource_b": {
+                "some_id": 2,
+                "region": "some_region",
+                "some_other_scope": "1"
+            }
+        },
+        "other_dummy_resources_count": 2
+    }
+}
+}
